+++ triangMemunc.py
@@ -24,10 +24,10 @@ print(' a       b             c')
 print('   x ----->')
 
 # co-ordinates input
-a,b,c = [int(x) for x in input("Enter the values for a,b and c: ").split()]
+a,b,c = [float(x) for x in input("Enter the values for a,b and c: ").split()]
 
 # x value input
-x=int(input("Enter the values for x: "))
+x=float(input("Enter the values for x: "))
 
 # Membership Function = y
